Handle byte constants in calls

+++ tests/concurrent_byte_index_or_assign/main.rs
@@ -150,11 +150,11 @@ impl<T> Iterator for GoChannel<T> {
 fn main() {
     let mut done = GoChannel::<bool>::new_buffered(1 as usize);
     let mut seen = Arc::new(Mutex::new(Some(vec![0; (2) as usize])));
-    let mut bit = Arc::new(Mutex::new(Some({ let __tmp_x = (*Arc::new(Mutex::new(Some(1 as u8))).lock().unwrap().as_ref().unwrap()); let __tmp_y = 3; __tmp_x << __tmp_y })));
+    let mut bit = Arc::new(Mutex::new(Some({ let __tmp_x = (*Arc::new(Mutex::new(Some(1 as u8))).lock().unwrap().as_ref().unwrap()); let __tmp_y = 3 as u8; __tmp_x << __tmp_y })));
 
     let bit_thread = Arc::new(Mutex::new(Some((*bit.lock().unwrap().as_ref().unwrap()).clone()))); let done_thread = done.clone(); let seen_thread = Arc::new(Mutex::new(Some((*seen.lock().unwrap().as_ref().unwrap()).clone()))); std::thread::spawn(move || {
         { let __idx = 0 as usize; let __rhs = (*bit.lock().unwrap().as_ref().unwrap()); let mut __seq_guard = seen_thread.lock().unwrap(); let __seq = __seq_guard.as_mut().unwrap(); __seq[__idx] = __seq[__idx] | __rhs; };;
-        done_thread.send({ let __tmp_x = { let __seq = { let __seq_holder = seen_thread.clone(); let __seq_guard = __seq_holder.lock().unwrap(); let __cloned = (*__seq_guard.as_ref().unwrap()).clone(); drop(__seq_guard); __cloned }; __seq[(0) as usize].clone() }; let __tmp_y = 8; __tmp_x == __tmp_y });;;
+        done_thread.send({ let __tmp_x = { let __seq = { let __seq_holder = seen_thread.clone(); let __seq_guard = __seq_holder.lock().unwrap(); let __cloned = (*__seq_guard.as_ref().unwrap()).clone(); drop(__seq_guard); __cloned }; __seq[(0) as usize].clone() }; let __tmp_y = 8 as u8; __tmp_x == __tmp_y });;;
     });
 
     println!("{}", done.recv().unwrap());
